refactor: Add token expiration configuration to YAML file
