+++ src/main/java/com/example/server/global/jwt/JwtAuthenticationFilter.java
@@ -46,42 +46,3 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		filterChain.doFilter(request, response);
 	}
 }
-
-// @Slf4j
-// @RequiredArgsConstructor
-// public class JwtAuthenticationFilter extends OncePerRequestFilter {
-//
-// 	private final AccessTokenUtil tokenProvider;
-//
-// 	@Override
-// 	protected void doFilterInternal(
-// 		HttpServletRequest request,
-// 		HttpServletResponse response,
-// 		FilterChain filterChain) throws ServletException, IOException {
-//
-// 		// 1. Request Header 에서 JWT 토큰 추출
-// 		final String authorizationHeader = request.getHeader("Authorization");
-//
-// 		// 2. validateToken 으로 토큰 유효성 검사
-// 		if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
-// 			String tokenStr = JwtHeaderUtil.getAccessToken(request);
-// 			AccessToken token = tokenProvider.convertAccessToken(tokenStr);
-// 			try {
-// 				if (token.validate()) {
-// 					Authentication authentication = tokenProvider.getAuthentication(token);
-// 					SecurityContextHolder.getContext().setAuthentication(authentication);
-// 				}
-// 			} catch (RedisConnectionFailureException e) {
-// 				SecurityContextHolder.clearContext();
-// 				e.printStackTrace();
-// 				//                throw new BaseException(REDIS_ERROR);
-// 			} catch (Exception e) {
-// 				e.printStackTrace();
-// 				//                throw new BaseException(INVALID_JWT);
-// 			}
-// 		} else {
-// 			SecurityContextHolder.getContext().setAuthentication(null);
-// 		}
-// 		filterChain.doFilter(request, response);
-// 	}
-// }
